backend/requirements.txt
# Web Framework
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI/ML
openai>=1.3.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Data
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
